feat: addressing code review issues
